pyproject.toml
[project]
name = "agent-community"
version = "0.1.0"
description = "Multi-agent collaboration community and communication bus."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "sqlalchemy>=2.0.32,<3.0.0",
  "asyncpg>=0.29.0,<1.0.0",
  "redis>=5.0.7,<6.0.0",
  "pydantic-settings>=2.4.0,<3.0.0",
  "httpx>=0.27.0,<1.0.0",
  "sse-starlette>=2.1.2,<3.0.0",
  "python-json-logger>=2.0.7,<3.0.0",
  "aiofiles>=24.1.0,<25.0.0",
  "email-validator>=2.2.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "pytest-asyncio>=0.23.8,<1.0.0",
  "ruff>=0.6.2,<1.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"
